requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
accelerate==1.2.1
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   starlette
    #   watchfiles
bitsandbytes==0.42.0
    # via -r requirements.in
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via uvicorn
fastapi==0.115.6
    # via -r requirements.in
filelock==3.16.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2024.9.0
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   torch
h11==0.14.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
huggingface-hub==0.27.0
    # via
    #   accelerate
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   requests
jinja2==3.1.4
    # via torch
loguru==0.7.3
    # via -r requirements.in
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==1.26.4
    # via
    #   -r requirements.in
    #   accelerate
    #   scipy
    #   transformers
packaging==24.2
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
psutil==6.1.0
    # via accelerate
pydantic==2.10.3
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.27.1
    # via pydantic
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0.2
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
    #   uvicorn
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   transformers
safetensors==0.4.5
    # via
    #   accelerate
    #   transformers
scipy==1.14.1
    # via bitsandbytes
sniffio==1.3.1
    # via anyio
starlette==0.41.3
    # via fastapi
sympy==1.13.3
    # via torch
tokenizers==0.21.0
    # via transformers
torch==2.2.2
    # via
    #   -r requirements.in
    #   accelerate
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.47.0
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
urllib3==2.2.3
    # via requests
uvicorn[standard]==0.34.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.3
    # via uvicorn
websockets==14.1
    # via uvicorn
